+ Added a new [td] tag which supports colspan and width attributes, I also added a couple unit tests for them as well. ! There was a problem that was not noticed until today which involved improper child tags. The parser would then fail to close the tags properly.
For example:
[table][tr][td undefinedattr=value]tests[/td][/tr][/table] 

would result in:
<table><tr>[td undefinedattr=value]tests[/td][/tr][/table] 

This was because once the interpreter got to the [td] tag with an undefined attribute it would not parse it, which would make it ignore that tag, its closing tag, and any other tags that depended on it. The problem was that the tags it depends on (and vice versa) may have been previously parsed (<table> and <tr>) so only their closing tags would then be ignored, so they would never be closed.

This was fixed by first closing all tags which have a parent dependency before calling setIgnore, so the above example will now result in:
<table><tr></tr></table>[td undefinedattr=value]tests[/td][/tr][/table]